+++ src/components/EmailControl/index.ts
@@ -1,0 +1,2 @@
+import EmailControl from './EmailControl';
+export { EmailControl };