+++ src/components/H1/H1.tsx
@@ -1,0 +1,14 @@
+import React, { FC } from 'react';
+import Typography from '@mui/material/Typography';
+import { TypographyProps } from '@mui/material';
+
+const H1: FC<TypographyProps> = (props) => (
+  <Typography
+    component="h1"
+    variant="h4"
+    sx={{ width: '100%', fontSize: 'clamp(2rem, 10vw, 2.15rem)' }}
+    {...props}
+  />
+);
+
+export default H1;