+++ src/components/PasswordControl/index.ts
@@ -1,0 +1,3 @@
+import PasswordControl from './PasswordControl';
+
+export { PasswordControl };