+++ src/components/RememberMeControl/index.ts
@@ -1,0 +1,3 @@
+import RememberMeControl from './RememberMeControl';
+
+export { RememberMeControl };