+++ src/components/TextField/TextField.tsx
@@ -1,0 +1,5 @@
+import React, { FC } from 'react';
+import MuiTextField, { TextFieldProps } from '@mui/material/TextField';
+
+const TextField: FC<TextFieldProps> = (props) => <MuiTextField {...props} />;
+export default TextField;